Revert "support Electron custom origins"
This reverts commit 7c2ba1ac2bf8c894fca632d8eefa5d4c3eaa0b31.
