+++ apps/desktop/src/app/DesktopClerk.test.ts
@@ -66,7 +66,7 @@ describe("DesktopClerk", () => {
           {
             storage: storageAdapter,
             passkeys: true,
-            renderer: { scheme: "t3code-dev", host: "app.t3.codes" },
+            renderer: { scheme: "t3code-dev", host: "app" },
           },
         ],
       ]);
@@ -139,7 +139,7 @@ describe("DesktopClerk", () => {
         {
           storage: storageAdapter,
           passkeys: true,
-          renderer: { scheme, host: "app.t3.codes" },
+          renderer: { scheme, host: "app" },
         },
       ],
     ]);
